Rename controller and routes
