+++ services/question/src/index.ts
@@ -7,16 +7,18 @@ if (process.env.NODE_ENV) {
 
 import express from 'express'
 import logger from './utils/logger'
-import router from './routes/questions'
+import router from './routes/questions-routes'
 import cors from 'cors'
 
 const app = express()
 const port = process.env.PORT || 3000
 
 app.use(express.json())
-app.use(cors({
-  origin: process.env.CORS_ORIGINS ? JSON.parse(process.env.CORS_ORIGINS) : '*'
-}))
+app.use(
+  cors({
+    origin: process.env.CORS_ORIGINS ? JSON.parse(process.env.CORS_ORIGINS) : '*',
+  })
+)
 
 app.use('/questions', router)
